p2p support

+++ build.ps1
@@ -203,7 +203,7 @@ function BuildCSproj()
     & $nugetExe restore -msbuildVersion 14 .\NuGet.Clients.sln
 
     # Build the solution
-    & $msbuildExe .\NuGet.Clients.sln "/p:Configuration=$Configuration;ReleaseLabel=$ReleaseLabel;BuildNumber=$BuildNumber;RunTests=$RunTests"
+    & $msbuildExe .\NuGet.Clients.sln /m "/p:Configuration=$Configuration;ReleaseLabel=$ReleaseLabel;BuildNumber=$BuildNumber;RunTests=$RunTests;BuildInParallel=true"
 
     if ($LASTEXITCODE -ne 0)
     {
@@ -285,16 +285,19 @@ if ((Test-Path $dnvmLoc) -eq $False)
 }
 
 ## Clean artifacts and nupkgs folder
-if (Test-Path $nupkgsDir)
+if ($Fast -eq $False)
 {
-    Trace-Log "Cleaning nupkgs folder"
-    Remove-Item $nupkgsDir\*.nupkg
-}
+    if (Test-Path $nupkgsDir)
+    {
+        Trace-Log "Cleaning nupkgs folder"
+        Remove-Item $nupkgsDir\*.nupkg
+    }
 
-if( Test-Path $artifacts)
-{
-    Trace-Log "Cleaning the artifacts folder"
-    Remove-Item $artifacts\*.* -Recurse
+    if( Test-Path $artifacts)
+    {
+        Trace-Log "Cleaning the artifacts folder"
+        Remove-Item $artifacts\*.* -Recurse
+    }
 }
 
 ## Make sure the needed DNX runtimes ex
